requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
drf-yasg==1.21.7
Pillow==10.1.0
python-decouple==3.8